Move creation event view from group to event app
